Move service infrastructure to dedicated module
